Refactoring class locations
