+++ src/main/java/com/sethhaskellcondie/thegamepensiveapi/controllers/HeartbeatController.java
@@ -1,4 +1,4 @@
-package com.sethhaskellcondie.thegamepensiveapi;
+package com.sethhaskellcondie.thegamepensiveapi.controllers;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
